% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker}
\alias{docker}
\title{Visualize Dockerfile}
\usage{
docker()
}
\value{
An interactive bar chart displaying Dockerfile instructions.
}
\description{
This function allows you to visualize a Dockerfile by extracting and displaying its instructions in a bar chart.
}
\examples{
\dontrun{
docker()
}

}
\seealso{
\code{\link{docker_compose}}, \code{\link{whats_docker}}, \code{\link{docker_info}}
}
\author{
Your Name
}
