% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_info.R
\name{docker_info}
\alias{docker_info}
\title{View Docker Information}
\usage{
docker_info()
}
\value{
An HTML widget displaying Docker information.
}
\description{
This function provides information about the Docker setup on your system, including whether Docker is installed,
if the Docker daemon is running, Docker version, all Docker images (only names), and all running Docker containers (only names).
}
\examples{
\dontrun{
docker_info()
}

}
\seealso{
\code{\link{docker_compose}}, \code{\link{whats_docker}}
}
\author{
Your Name
}
