% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubernetes_info.R
\name{kubernetes_info}
\alias{kubernetes_info}
\title{Visualize Kubernetes Information}
\usage{
kubernetes_info()
}
\value{
An interactive table displaying Kubernetes cluster information if Kubernetes is installed.
}
\description{
This function checks if Kubernetes is installed on the system and displays relevant information
about the Kubernetes cluster if it's installed.
}
\examples{
\dontrun{
kubernetes_info()
}

}
\seealso{
\code{\link{docker_compose}}, \code{\link{whats_docker}}, \code{\link{docker_info}}, \code{\link{docker}}
}
\author{
Your Name
}
