% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whats_docker.R
\name{whats_docker}
\alias{whats_docker}
\title{What is Docker?}
\usage{
whats_docker()
}
\value{
An HTML widget with information about Docker.
}
\description{
Docker is an open-source platform that uses containerization technology to enable developers
to package applications and their dependencies into a virtual container. This makes it possible
to run applications consistently across various computing environments.
}
\examples{
\dontrun{
whats_docker()
}

}
\seealso{
\code{\link{docker_info}}, \code{\link{docker_compose}}

https://www.docker.com/

https://docs.docker.com/

https://docs.docker.com/get-started/

https://hub.docker.com/

https://www.docker.com/play-with-docker
}
\author{
Matt Majestic
}
